fix: verse cache revalidation and countdown
The server was returning a different current time, and the countdown was being cached on the CDNs, so the user could never see the real time for a new verse, resulting in inaccurate countdown calculations. Additionally, the cache revalidation strategy was not functioning correctly.

we've made some changes:

We replaced the date-fns library with Luxon for easier and more precise date and timezone manipulation. Additionally, we simplified the verse cache revalidation strategy: instead of relying on a time-based approach, we've implemented an on-demand revalidation. To achieve this, we added a Vercel cron job to trigger the revalidation of the verse every day at 6 AM Brasília time.

Also forced the route to be dynamic to prevent it from being cached on the CDNs.

These changes aim to improve the accuracy of time calculations and ensure effective cache revalidation, providing a more consistent and reliable user experience. If anyone has a better idea or improvement to this implementation, pull requests and suggestion are highly welcome!