refactor: use useMemo for `countdown` optimization
I used the useMemo hook to optimize the performance of the component, caching calculated values between re-renders. Although this optimization doesn't bring a visually perceptible difference in performance, it's noticeable when observing the numbers, especially when using tools like the React Developer Tools extension. It was an interesting opportunity to explore and better understand this React hook.